# docs/swagger.yaml
definitions:
...
  models.AutuhorityInfo:
    properties:
      createDate:
        type: string
      creator:
        type: string
      id:
        type: string
      method:
        type: string
      modifier:
        type: string
      modifyDate:
        type: string
      name:
        type: string
      realm:
        type: string
      url:
        type: string
      useYn:
        type: boolean
    required:
    - id
    - name
    type: object
...
  models.Id:
    properties:
      id:
        type: string
    required:
    - id
    type: object
...
  models.MenuAutuhorityInfo:
    properties:
      method:
        type: string
      name:
        type: string
      url:
        type: string
    required:
    - name
    type: object
...
  models.RolesInfo:
    properties:
      authId:
        items:
          type: string
        type: array
      createDate:
        type: string
      creator:
        type: string
      defaultRole:
        type: boolean
      id:
        type: string
      modifier:
        type: string
      modifyDate:
        type: string
      name:
        type: string
      realm:
        type: string
      tenantId:
        type: string
      useYn:
        type: boolean
    required:
    - id
    - name
    type: object
...
info:
  contact: {}
  title: IAM.Backend API Document
  version: "1.0"
paths:
...
  /authority/{tenantId}/group/{groupId}/roles/{roleId}:
    delete:
      parameters:
      - description: Realm Id
        in: path
        name: realm
        required: true
        type: string
      - description: tenantId
        in: query
        name: tenantId
        required: true
        type: string
      - description: Group Id
        in: path
        name: groupId
        required: true
        type: string
      - description: Role Id
        in: path
        name: roleId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: 그룹 역할 제외
      tags:
      - Authority
  /authority/{tenantId}/user/{userId}/roles/{roleId}:
    delete:
      parameters:
      - description: Realm Id
        in: path
        name: realm
        required: true
        type: string
      - description: tenantId
        in: query
        name: tenantId
        required: true
        type: string
      - description: User Id
        in: path
        name: userId
        required: true
        type: string
      - description: Role Id
        in: path
        name: roleId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: 유저 역할 제외
      tags:
      - Authority
  /authority/auth:
    get:
      parameters:
      - description: Realm Id
        in: path
        name: realm
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.AutuhorityInfo'
            type: array
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: 전체 권한 목록 조회
      tags:
      - Authority
    post:
      parameters:
      - description: Realm Id
        in: path
        name: realm
        required: true
        type: string
      - description: body
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/models.AutuhorityInfo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Id'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: 권한 생성
      tags:
      - Authority
  /authority/auth/{authId}:
    delete:
      parameters:
      - description: Realm Id
        in: path
        name: realm
        required: true
        type: string
      - description: Auth Name
        in: path
        name: authId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Id'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: 권한 삭제
      tags:
      - Authority
    get:
      parameters:
      - description: Auth Id
        in: path
        name: authId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AutuhorityInfo'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: 권한 정보 조회
      tags:
      - Authority
    put:
      parameters:
      - description: Realm Id
        in: path
        name: realm
        required: true
        type: string
      - description: Auth Name
        in: path
        name: authId
        required: true
        type: string
      - description: body
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/models.AutuhorityInfo'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: 권한 수정
      tags:
      - Authority
  /authority/auth/menu/{site}:
    get:
      parameters:
      - description: Site
        in: path
        name: site
        required: true
        type: string
      - description: tenantId
        in: path
        name: tenantId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.MenuAutuhorityInfo'
            type: array
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: 메뉴 권한 조회
      tags:
      - Authority
  /authority/group/{groupId}/roles:
    post:
      parameters:
      - description: tenantId
        in: query
        name: tenantId
        required: true
        type: string
      - description: Group Id
        in: path
        name: groupId
        required: true
        type: string
      - description: Role Id
        in: body
        name: roleId
        required: true
        schema:
          $ref: '#/definitions/models.Id'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: 그룹 역할 할당
      tags:
      - Authority
  /authority/roles:
    get:
      parameters:
      - description: 사용되지 않음
        in: path
        name: realm
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.RolesInfo'
            type: array
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: 전체 역할 목록 조회
      tags:
      - Authority
    post:
      parameters:
      - description: body
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/models.RolesInfo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Id'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: 역할 생성
      tags:
      - Authority
  /authority/roles/{roleId}:
    delete:
      parameters:
      - description: Role Id
        in: path
        name: roleId
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: 역할 삭제
      tags:
      - Authority
    put:
      parameters:
      - description: Role Id
        in: path
        name: roleId
        required: true
        type: string
      - description: body
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/models.RolesInfo'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: 역할 수정
      tags:
      - Authority
  /authority/roles/{roleId}/auth:
    get:
      parameters:
      - description: Role Id
        in: path
        name: roleId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.RolesInfo'
            type: array
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: 역할 할당 권한 목록 조회
      tags:
      - Authority
    post:
      parameters:
      - description: Role Id
        in: path
        name: roleId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: 역할 권한 할당
      tags:
      - Authority
  /authority/roles/{roleId}/auth/{authId}:
    delete:
      parameters:
      - description: Role Id
        in: path
        name: roleId
        required: true
        type: string
      - description: Auth Id
        in: path
        name: authId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: 역할 권한 제외
      tags:
      - Authority
    put:
      parameters:
      - description: Role Id
        in: path
        name: roleId
        required: true
        type: string
      - description: Auth Id
        in: path
        name: authId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: 역할 권한 수정
      tags:
      - Authority
  /authority/user/{userId}:
    get:
      parameters:
      - description: Realm Id
        in: path
        name: realm
        required: true
        type: string
      - description: User Id
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.RolesInfo'
            type: array
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: 유저 할당 역할 목록 조회
      tags:
      - Authority
...
  /authority/user/auth:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: 유저 권한 조회
      tags:
      - Authority
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"